First Post is ready!!!!
Finished creating the koi fish on the lock screen. Created the very first post and finished it. Yippeee!

Will look into upgrading the website and getting an actual domain. Look into hosting on one of my servers and securing it. Look into using a CMS and dynamic blog creation. Many upgrades to look into the future.